+++ package.json
@@ -50,6 +50,7 @@
     "eslint-plugin-standard": "^4.0.1",
     "git-pre-hooks": "^1.2.1",
     "jest": "^25.5.1",
+    "mock-spawn": "^0.2.6",
     "prettier": "^2.0.5"
   },
   "jest": {
